requirements.txt
python-telegram-bot==21.6
requests
icalendar
python-dotenv
